Copy function `guid` to context-bound version
jQuery internally relies on unique `guid` values to ensure that the
correct function is removed from an element in `$(element).off('event',
func);`. Previously, `component.on` would attach the `guid` value to the
original callback to preserve this functionality in jQuery. However,
future calls to `.on('event', sameFunc)` would fail to check the
original callback for this `guid` value and simply overwrite it with
whatever new `guid` was assigned by jQuery. This change preserves the
original `guid` value across multiple calls to `on` with the same
function.