Updated .trigger() to parse the three argument case.
https://github.com/flightjs/flight/blob/master/doc/base_api.md#thistriggerselector-eventtype--eventpayload
seems to indicate that passing three arguments into this.trigger(...) is
a supported case and should parse the arguments in the following order:

1. selector
2. eventType
3. eventPayload

However, I ran into issues when delegating the event trigger (e.g. to
the document object) with an event payload.

This commit includes a unit test exposing the problematic case and a
patch to base.js to properly parse the argument list for it.